Move old TCP analyzer into analyzer adapter in packet analysis tree
