+++ src/analyzer/Manager.cc
@@ -81,14 +81,6 @@ Manager::~Manager()
 		}
 	}
 
-void Manager::InitPreScript()
-	{
-	// Cache these tags.
-	analyzer_connsize = GetComponentTag("CONNSIZE");
-	analyzer_stepping = GetComponentTag("STEPPINGSTONE");
-	analyzer_tcpstats = GetComponentTag("TCPSTATS");
-	}
-
 void Manager::InitPostScript()
 	{
 	const auto& id = detail::global_scope()->Find("Tunnel::vxlan_ports");
@@ -352,132 +344,6 @@ Manager::tag_set* Manager::LookupPort(TransportProto proto, uint32_t port, bool 
 	tag_set* l = new tag_set;
 	m->insert(std::make_pair(port, l));
 	return l;
-	}
-
-bool Manager::BuildInitialAnalyzerTree(Connection* conn)
-	{
-	analyzer::tcp::TCP_Analyzer* tcp = nullptr;
-	packet_analysis::IP::SessionAdapter* root = nullptr;
-	analyzer::pia::PIA* pia = nullptr;
-	bool check_port = false;
-
-	switch ( conn->ConnTransport() ) {
-
-	case TRANSPORT_TCP:
-		root = tcp = new analyzer::tcp::TCP_Analyzer(conn);
-		pia = new analyzer::pia::PIA_TCP(conn);
-		check_port = true;
-		DBG_ANALYZER(conn, "activated TCP analyzer");
-		break;
-
-	default:
-		reporter->InternalWarning("unknown protocol can't build analyzer tree");
-		return false;
-	}
-
-	bool scheduled = ApplyScheduledAnalyzers(conn, false, root);
-
-	// Hmm... Do we want *just* the expected analyzer, or all
-	// other potential analyzers as well?  For now we only take
-	// the scheduled ones.
-	if ( ! scheduled )
-		{ // Let's see if it's a port we know.
-		if ( check_port && ! zeek::detail::dpd_ignore_ports )
-			{
-			int resp_port = ntohs(conn->RespPort());
-			tag_set* ports = LookupPort(conn->ConnTransport(), resp_port, false);
-
-			if ( ports )
-				{
-				for ( tag_set::const_iterator j = ports->begin(); j != ports->end(); ++j )
-					{
-					Analyzer* analyzer = analyzer_mgr->InstantiateAnalyzer(*j, conn);
-
-					if ( ! analyzer )
-						continue;
-
-					root->AddChildAnalyzer(analyzer, false);
-					DBG_ANALYZER_ARGS(conn, "activated %s analyzer due to port %d",
-					                  analyzer_mgr->GetComponentName(*j).c_str(), resp_port);
-					}
-				}
-			}
-		}
-
-	if ( tcp )
-		{
-		// We have to decide whether to reassamble the stream.
-		// We turn it on right away if we already have an app-layer
-		// analyzer, reassemble_first_packets is true, or the user
-		// asks us to do so.  In all other cases, reassembly may
-		// be turned on later by the TCP PIA.
-
-		bool reass = root->GetChildren().size() ||
-				zeek::detail::dpd_reassemble_first_packets ||
-				zeek::detail::tcp_content_deliver_all_orig ||
-				zeek::detail::tcp_content_deliver_all_resp;
-
-		if ( tcp_contents && ! reass )
-			{
-			static auto tcp_content_delivery_ports_orig = id::find_val<TableVal>("tcp_content_delivery_ports_orig");
-			static auto tcp_content_delivery_ports_resp = id::find_val<TableVal>("tcp_content_delivery_ports_resp");
-			const auto& dport = val_mgr->Port(ntohs(conn->RespPort()), TRANSPORT_TCP);
-
-			if ( ! reass )
-				reass = (bool)tcp_content_delivery_ports_orig->FindOrDefault(dport);
-
-			if ( ! reass )
-				reass = (bool)tcp_content_delivery_ports_resp->FindOrDefault(dport);
-			}
-
-		if ( reass )
-			tcp->EnableReassembly();
-
-		if ( IsEnabled(analyzer_stepping) )
-			{
-			// Add a SteppingStone analyzer if requested.  The port
-			// should really not be hardcoded here, but as it can
-			// handle non-reassembled data, it doesn't really fit into
-			// our general framing ...  Better would be to turn it
-			// on *after* we discover we have interactive traffic.
-			uint16_t resp_port = ntohs(conn->RespPort());
-			if ( resp_port == 22 || resp_port == 23 || resp_port == 513 )
-				{
-				static auto stp_skip_src = id::find_val<TableVal>("stp_skip_src");
-				auto src = make_intrusive<AddrVal>(conn->OrigAddr());
-
-				if ( ! stp_skip_src->FindOrDefault(src) )
-					tcp->AddChildAnalyzer(new analyzer::stepping_stone::SteppingStone_Analyzer(conn), false);
-				}
-			}
-
-		if ( IsEnabled(analyzer_tcpstats) )
-			// Add TCPStats analyzer. This needs to see packets so
-			// we cannot add it as a normal child.
-			tcp->AddChildPacketAnalyzer(new analyzer::tcp::TCPStats_Analyzer(conn));
-
-		if ( IsEnabled(analyzer_connsize) )
-			// Add ConnSize analyzer. Needs to see packets, not stream.
-			tcp->AddChildPacketAnalyzer(new analyzer::conn_size::ConnSize_Analyzer(conn));
-		}
-
-	else
-		{
-		if ( IsEnabled(analyzer_connsize) )
-			// Add ConnSize analyzer. Needs to see packets, not stream.
-			root->AddChildAnalyzer(new analyzer::conn_size::ConnSize_Analyzer(conn));
-		}
-
-	if ( pia )
-		root->AddChildAnalyzer(pia->AsAnalyzer());
-
-	conn->SetSessionAdapter(root, pia);
-	root->Init();
-	root->InitChildren();
-
-	PLUGIN_HOOK_VOID(HOOK_SETUP_ANALYZER_TREE, HookSetupAnalyzerTree(conn));
-
-	return true;
 	}
 
 void Manager::ExpireScheduledAnalyzers()
