+++ src/fuzzers/pop3-fuzzer.cc
@@ -36,7 +36,7 @@ static zeek::Connection* add_connection()
 
 static zeek::analyzer::Analyzer* add_analyzer(zeek::Connection* conn)
 	{
-	auto* tcp = new zeek::analyzer::tcp::TCP_Analyzer(conn);
+	auto* tcp = new zeek::packet_analysis::TCP::TCPSessionAdapter(conn);
 	auto* pia = new zeek::analyzer::pia::PIA_TCP(conn);
 	auto a = zeek::analyzer_mgr->InstantiateAnalyzer(ZEEK_FUZZ_ANALYZER, conn);
 	tcp->AddChildAnalyzer(a);
