+++ src/packet_analysis/protocol/tcp/Plugin.cc
@@ -3,6 +3,7 @@
 #include "zeek/plugin/Plugin.h"
 #include "zeek/packet_analysis/Component.h"
 #include "zeek/packet_analysis/protocol/tcp/TCP.h"
+#include "zeek/packet_analysis/protocol/tcp/TCPSessionAdapter.h"
 
 namespace zeek::plugin::Zeek_TCP {
 
@@ -12,6 +13,8 @@ public:
 		{
 		AddComponent(new zeek::packet_analysis::Component("TCP",
 		                 zeek::packet_analysis::TCP::TCPAnalyzer::Instantiate));
+		AddComponent(new zeek::analyzer::Component("TCP",
+		                 zeek::packet_analysis::TCP::TCPSessionAdapter::Instantiate));
 
 		zeek::plugin::Configuration config;
 		config.name = "Zeek::TCP_PKT";
